Lab 5 Answers:
3.
a. Because Java will already have a constructor (implicit) , especially if the programmer
does not write one themselves. Java creates a constuctor already (behind the scenes) if a no argument
constructor is implimented; java will then invoke this constructor it has made. Inhereitance is
part of this too. Each class can extrend anothers, thus this particular class does'nt need
a implicit constructor. LinkedStack inherites from IStack and extends it.

b. It is the total stack space that has been used, in our case however many we add.
It is the extra space or temporary space used by that algorithm, the one that creates the stack
and adds to it.

c. You could create a list rotation method. If it is negative the list/array that you want to
re-create would move each element to the left. As long as the count/number of
rotations is less than 0, you could add to the count by list.size();
Also you would later create a new LinkedList (maybe?) and do Arrays.asList then
do a while loop to execute while loop is true and rotate the list.

d. For me, personally, I had more success with pushing in stacks. That way you
dont have to worry about deleting an element from the stack. If I was using pop,
I would do an if(isEmpty) throw new NoSuchElementException ....
then have the first node of the stack be linked to the item and then first.next;
Pop just has a more moving parts to it than a simple push.