Question #1: Why does LinkedStack not require an explicit constructor? Because LinkedStack does not require an explicit constructor. This is because you have methods already created in its class that can be called. Because of this, it eliminates the use/need for a constructor.
What is the time and (extra) space complexity of each of the LinkedStack methods, as well as ReverseLines.main?
For LinkedStack methods the push, peek, and pop methods have a complexity of T(n) = O(1), and a space complexity of S(n) = 0(1); in constant time. Then,for the main class in ReverseLines the isEmpty and asList both have a time complexity of T(n) = O(1) and S(n) = O(1); in constant time.

How else (not using Node) could we have implemented LinkedStack in such a way that it is still based on a linked list but the asList method uses constant time and space?
You maybe could have done that by using an arrayList as the container so you would have O(n).

Is it better for push and pop to return the item or the stack itself? Briefly discuss the pros and cons of each design.
Pros:

Also, operating on constant time and space is a plus.
Using push and pop methods will use less memory, in turn, making them faster.
Cons:

Returning the entire stack at once means that every value is returned at once vs. individually.
Returning a large stack will usually take more time.